refactor: create fixtures utility
