+++ test/core.test.ts
@@ -1,23 +1,10 @@
-import { exec } from 'child_process';
 import { dirname, join } from 'path';
 import { fileURLToPath } from 'url';
-import { promisify } from 'util';
 import { ESLint, Linter } from 'eslint';
 import { Core } from '../src/core.js';
+import { cleanupFixturesCopy, getSnapshotOfChangedFiles, setupFixturesCopy } from './test-util/fixtures.js';
 
-const execPromise = promisify(exec);
 const cwd = join(dirname(fileURLToPath(import.meta.url)), '..');
-
-/**
- * Returns a string containing the stitched together contents of the file modified by transform.
- * To make the snapshot easier to read, the name of the file is inserted at the beginning of the contents of each file.
- */
-async function getSnapshotOfChangedFiles(): Promise<string> {
-  const { stdout } = await execPromise(`diff -qr fixtures fixtures-tmp | cut -d " " -f 4 | xargs tail -n +1`, {
-    cwd,
-  });
-  return stdout.toString();
-}
 
 // Normalize `message` for the snapshot.
 function normalizeMessage(message: Linter.LintMessage): Linter.LintMessage {
@@ -46,11 +33,11 @@ function normalizeResults(results: ESLint.LintResult[]): ESLint.LintResult[] {
 }
 
 beforeEach(async () => {
-  await execPromise(`rm -rf fixtures-tmp && cp -r fixtures fixtures-tmp`, { cwd });
+  await setupFixturesCopy();
 });
 
 afterEach(async () => {
-  await execPromise(`rm -rf fixtures-tmp`, { cwd });
+  await cleanupFixturesCopy();
 });
 
 describe('Core', () => {
